requirements.txt
python-dotenv>=1.0.1
SQLAlchemy>=1.4.0
pydantic>=2.0.0
aiosqlite>=0.19.0
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0
python-dateutil>=2.8.0
